feat(api): add config.go file to handle API configuration
This commit adds a new file `config.go` to the `api` package. This file contains functions and constants related to API configuration. The `GetGenerateFunction` function is added to retrieve the appropriate generate function based on the API type specified in the configuration. The function takes into account the `model` and `API_TYPE` configuration values to create the appropriate LLMS instance and return the corresponding generate function.

feat(cmd): add API_TYPE flag to root command

This commit adds the `API_TYPE` flag to the root command. The flag allows the user to specify the API type to be used for generating content. The default value is set to `API_OLLAMA`. The flag is registered with a completion function that provides suggestions for the available API types.

fix(prompt): change keybinding for chat update

This commit changes the keybinding for chat update from `CtrlF` to `CtrlI`. The previous keybinding was conflicting with other keybindings and causing issues. The new keybinding should not conflict with any existing keybindings.

fix(prompt): handle remove stack event properly

This commit fixes a bug in the `chatModel` where the remove stack event was not being handled properly. The code now checks if the stack is not nil before trying to find the index of the stack in the stack slice. This prevents a panic when comparing the stack types.

fix(prompt): handle error in sendPrompt function

This commit adds error handling to the `sendPrompt` function in the `chatUpdate.go` file. The function now returns an error if there is an error in adding the user message or assistant message to the chat messages.