chore(chat): remove unused errMsg type
The errMsg type was defined but not used anywhere in the code. Removing it to improve code cleanliness.

feat(chat): add loading state to chatModel

Added a new boolean field `loading` to the `chatModel` struct to keep track of whether the chat is in a loading state or not. This will be used to change the background color of the title based on the loading state.

fix(chat): fix transition behavior in Update function

Fixed the transition behavior in the `Update` function of the `chatModel`. Previously, the `transition` flag was being set to `false` unconditionally, which caused incorrect behavior. Removed the unnecessary code.

fix(chat): fix promptSend function

Fixed the `promptSend` function in the `chatUpdate.go` file. The error handling code was commented out, which caused the error to be ignored. Uncommented the error handling code to properly handle errors.

feat(style): add loading and normal background colors

Added two new constants `LoadingBackgroundColor` and `NormalBackgroundColor` to the `style` package. These colors will be used to set the background color of the title based on the loading state in the `chatModel`.